+++ src/client/controllers/board_controller.js
@@ -3,7 +3,6 @@ var BoardCreator = require('../board_creator');
 var UserMod = require('../user_mod');
 
 module.exports = ['$http', function($http) {
-  app.board = this; // for debug
   var board = this;
   this.projectLinker = new ProjectLinker(board, $http);
   this.userMod = new UserMod(board, $http);
@@ -17,6 +16,7 @@ module.exports = ['$http', function($http) {
     }
   };
   this.load = app.loadBoard = function (attributes) {
+    app.board = this; // for debug
     board.creator.isOpen = false;
     board.attributes = attributes;
     board.loaded = true;
@@ -41,7 +41,7 @@ module.exports = ['$http', function($http) {
     }
   }
   this.removeCard = function(col, row) {
-    if (confirm("Are you sure you wish to delete this card?")) {
+    if (col > -1 && row > -1 && confirm("Are you sure you wish to delete this card?")) {
       $http.delete(api.route('boards/'+board.attributes._id+'/columns/'+col+'/cards/'+row)).success(function(data) {
         if (data.board)
           board.attributes.columns = data.board.columns;
